BK64: Move the geo layout into its own module
